Make script generate a fixed point
When called with

    kpt fn run .

the script in the container image will receive a functionConfig with
the configmap. If it fails to output that configmap, the file will be
removed. So the script must include it in the output, making sure it
doesn't mutate it unnecessarily.

There's another couple of modes the script can be run in:

 - on its own, or piping `kpt fn source .` into it;
 - with `kpt fn run . --image -- item=value`

For the first form, the script constructs a default `functionConfig`.
For the second, it accepts the `functionConfig` it gets, but renames
it so it'll be put in the expected file.

This should mean that running `kpt fn run .` without changing anything
will not make changes to any files.