Separate functionConfig and instance files
To be able to parameterise the functionConfig with `kpt cfg`, I need
the functionConfig resource to _not_ be round-tripped though `kpt fn`,
since doing so will in general (and certainly in this case) strip the
comments used by `kpt cfg`.

This moves the functionConfig to its own directory; and, since `kpt
fn` will return early when ther are no resources, adds the output to
instance/ so it'll run. It now needs to be run with

    kpt fn run --fn-path ./fn/ ./instance/

since running in the current directory would end up pruning fn/.