Add a setter for the image argument
This adds a setter for the _functionConfig_ field `image`, with the
idea being that it provides the API for the package.

Note that I first changed the value of the image field to be different
to what was in the instance file, so that when the setter was created,
it didn't get attached to the latter. Then I ran

    kpt fn run --fn-path=fn ./instance

to regenerate the instance files.